star-height: minor changes
* lib/vcsn/algos/derivation.cc (star_height): Move to...
* lib/vcsn/algos/star-height.cc: here.
* lib/vcsn/local.mk: Adjust.

* vcsn/algos/derivation.hh: Remove now useless include.

* vcsn/algos/star-height.hh, vcsn/algos/star-height.hxx: Use 'unsigned'
rather than 'size_t', since we have provided explicit conversion rules
for the former, and not the latter.  It is also more consistent with
the other algorithms, that do use unsigned.
* vcsn/dyn/algos.hh, python/vcsn_python.cc: Adjust.

+++ python/vcsn_python.cc
@@ -341,7 +341,7 @@ struct ratexp
     return vcsn::dyn::standard(r_);
   }
 
-  size_t star_height() const
+  unsigned star_height() const
   {
     return vcsn::dyn::star_height(r_);
   }
